backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
starlette
gitpython
requests
PyGithub
pathspec
